+++ src/coastlines.jl
@@ -1,14 +1,7 @@
-"""
-    CoastLines
-"""
-struct CoastLines
-    resolution::Int
-    raw_points::Vector{Vector{POINT_LATLON{Float32}}}
-end
-
 const COASTLINES_DICT = Dict{Int, CoastLines}()
 
-function get_coastlines(; resolution = 110, force = false)
+function get_coastlines(; resolution = nothing, force = false)
+    resolution = check_resolution(resolution; force)
     force && haskey(COASTLINES_DICT, resolution) && delete!(COASTLINES_DICT, resolution)
     get!(COASTLINES_DICT, resolution) do
         ne_data = naturalearth("coastline", resolution)